requirements.txt
streamlit~=1.45.0
paho-mqtt~=2.1.0
pandas~=2.2.3
numpy~=2.2.2
requests~=2.32.3
ultralytics~=8.3.131